+++ test/devices_tests/switch_test.py
@@ -36,7 +36,9 @@ class TestSwitch(unittest.TestCase):
     def test_sync(self):
         """Test sync function / sending group reads to KNX bus."""
         xknx = XKNX()
-        switch = Switch(xknx, "TestOutlet", group_address_state="1/2/3")
+        switch = Switch(
+            xknx, "TestOutlet", group_address_state="1/2/3", group_address="1/2/4"
+        )
         self.loop.run_until_complete(switch.sync())
 
         self.assertEqual(xknx.telegrams.qsize(), 1)
@@ -392,3 +394,9 @@ class TestSwitch(unittest.TestCase):
         self.assertEqual(switch1.state, False)
         callback_mock.assert_called_once()
         callback_mock.reset_mock()
+
+    def test_unique_id(self):
+        """Test unique id functionality."""
+        xknx = XKNX()
+        switch = Switch(xknx, "TestOutlet", group_address="1/2/3")
+        self.assertEqual(switch.unique_id, "1/2/3")